Camera: Build android.hardware.camera.common@1.0-helper and replace libcamera_client
Our old camera HALs link libcamera_client, which isn't exposed to vendor, so we
need to instead link against what was supposed to replace libcamera_client:
android.hardware.camera.common@1.0-helper. However, this was supposed to be
statically linked into the camera HAL, so work around this by building a shared
version of it and then replacing the libcamera_client link in the camera HAL with it.

Change-Id: Ica7220222509dc0a8db86f73d338d605deb30165